requirements.txt
kivy==2.1.0
pyrebase4==4.5.0
requests==2.31.0
